Clarify what is interpretted as trust anchor for chain
This standardizes client behavior on considering the last certificate in
the chain as a trust anchor, regardless if it's a root or intermediate.
This means clients should use all preceding CA certificates as
chain-builders or intermediates.

We do document the case where an intermediate may be issued from
multiple roots, specifying that multiple chains should be provided.
Another solution for a later PR could be to provide two pools, trusted
and untrusted, as an alternative to chains.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>