Introduce v0.2 TrustedRoot, un-deprecate log ID
As we learned when we introduced a new SigningConfig version, adding
fields to the proto isn't a breaking change, but setting these fields
in JSON and unmarshalling the object is a breaking change. Since we've
added operator and checkpoint_key_id to the Instance message, this
requires introduced a new TrustedRoot version, which I've done in this
PR as v0.2. As with SigningConfig, I've made it optional for clients to
indefinitely support old versions, but it should be little work for a
client to support both versions concurrently.

I've also un-deprecated log ID. This was premature since no clients
supported checkpoint key ID yet. With Rekor v2, we'll be setting the
same value in both log_id and checkpoint_key_id, so clients can continue
to ignore checkpoint_key_id for now.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>