Relaxed the requirement of only one instance allowed to have open timespan. (#732)
This requirement causes problems with running multiple instances as can
be specified in the signing config. For a client's perspective, this does
not really matter as not having an end-time is similar to having an end
time that's set in the _far future_. As clients must work with overlaps
from instances, this should not be a breaking change.

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>